Essaie for very high precision.
Tried to use fraction instead of floats. Worked mostely except in the
shader where loops are done with workarround and therefore the modulo
and gcd(greatest common divisor) function aren't precise anymore. But I
keep the branch because of the cool giltch effect you get when you start
to zoom ~10 times